% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BCL3}
\alias{BCL3}
\title{Breast cancer biopsy layer 3 bulb spatial transcriptomics transcriptomics}
\format{List where 'counts' is a sparse matrix with columns as voxels and rows as genes and
                   'pos' is a data frame of x and y position values per voxel}
\source{
\url{http://www.spatialtranscriptomicsresearch.org/datasets/doi-10-1126science-aaf2403/}
}
\usage{
BCL3
}
\description{
Breast cancer biopsy layer 3 bulb spatial transcriptomics transcriptomics
}
\keyword{datasets}
