% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cd}
\alias{cd}
\title{MOB spatial transcriptomics transcriptomics}
\format{Sparse matrix with columns as voxels and rows as genes}
\source{
\url{http://www.spatialtranscriptomicsresearch.org/datasets/doi-10-1126science-aaf2403/}
}
\usage{
cd
}
\description{
MOB spatial transcriptomics transcriptomics
}
\keyword{datasets}
