% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drosophila}
\alias{drosophila}
\title{Drosophila embryo aligned ISH from the in situ database (BDTNP).}
\format{
List where 'mat' is a matrix with columns as cells and rows as genes and
                   'pos' is a data frame of x, y, and z position values per cells
}
\source{
\url{https://shiny.mdc-berlin.de/DVEX/}
}
\usage{
drosophila
}
\description{
Drosophila embryo aligned ISH from the in situ database (BDTNP).
}
\keyword{datasets}
