% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filterSpatialPatterns}
\alias{filterSpatialPatterns}
\title{Filter for spatial patterns}
\usage{
filterSpatialPatterns(mat, I, adjustPv = TRUE, alpha = 0.05,
  minPercentCells = 0.05, verbose = TRUE)
}
\arguments{
\item{mat}{Gene expression matrix.}

\item{I}{Output of getSpatialPatterns}

\item{alpha}{P-value threshold for LISA score to be considered significant.}

\item{minPercentCells}{Minimum percent of cells that must be driving spatial pattern}
}
\description{
Filter out spatial patterns driven by small number of cells using LISA
}
