% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestNeighbor.R
\name{getCrossLayerNeighbors}
\alias{getCrossLayerNeighbors}
\title{Nearest neighbors across layers}
\usage{
getCrossLayerNeighbors(layers, k = 3)
}
\arguments{
\item{layers}{List of positions}

\item{k}{Number of nearest neighbors}
}
\value{
Boolean matrix where value = 1 if two cells are considered adjacency ie. neighbors, else 0
}
\description{
Nearest neighbors across layers
}
