% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getDensity}
\alias{getDensity}
\title{Get density of points in 2 dimensions.}
\usage{
getDensity(pos, n = 100)
}
\arguments{
\item{n}{Create a square n by n grid to compute density.}

\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
The density within each square.
}
\description{
Get density of points in 2 dimensions.
}
\author{
Kamil Slowikowski http://slowkow.com/notes/ggplot2-color-by-density/
}
