% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestNeighbor.R
\name{getInterCellTypeWeight}
\alias{getInterCellTypeWeight}
\title{Filter adjacency weight matrix to between two subsets of points}
\usage{
getInterCellTypeWeight(cct, nct, weight, pos = NULL, plot = FALSE, ...)
}
\arguments{
\item{cct}{Cells of cell-type 1}

\item{nct}{Cells of cell-type 2 (assumed to be mutually exclusive with cct)}

\item{weight}{Adjacency weight matrix}

\item{pos}{Position}

\item{plot}{Boolean of whether to plot}

\item{...}{Additional plotting parameters}
}
\value{
Boolean matrix where value = 1 if two cells are considered adjacency ie. neighbors, else 0
}
\description{
Restrict adjacency relationships to between two subsets of points
}
