% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationStatistics
%   (client-10-245-74-74.fas.harvard.edu's conflicted copy 2019-11-09).R,
%   R/correlationStatistics.R
\name{getPv}
\alias{getPv}
\title{Helper function to get p-value from value and distribution mean and standard deviation}
\usage{
getPv(obs, ei, sdi, alternative)

getPv(obs, ei, sdi, alternative)
}
\arguments{
\item{obs}{Observed value}

\item{ei}{Mean (Expected value)}

\item{sdi}{Standard deviation}

\item{alternative}{"two.sided", "less", or "greater"}

\item{obs}{Observed value}

\item{ei}{Mean (Expected value)}

\item{sdi}{Standard deviation}

\item{alternative}{"two.sided", "less", or "greater"}
}
\value{
P-value

P-value
}
\description{
Helper function to get p-value from value and distribution mean and standard deviation

Helper function to get p-value from value and distribution mean and standard deviation
}
\examples{
# Probability of observing 5 while expecting 0 with a standard deviation of 1
getPv(5, 0, 1, 'greater')

# Probability of observing 5 while expecting 0 with a standard deviation of 1
getPv(5, 0, 1, 'greater')

}
