% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getSpatialWeights}
\alias{getSpatialWeights}
\title{Derive spatial weights}
\usage{
getSpatialWeights(pos, klist = 3:9, plot = FALSE)
}
\arguments{
\item{pos}{Position matrix where each row is a cell, columns are
x, y, (optionally z) coordinations}

\item{klist}{range of number of nearest neighbors to consider (default 3:9)}

\item{plot}{Whether to plot neighbor network}
}
\description{
Derive spatial weights given position of cells in space
}
