% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationStatistics.R
\name{lisaTest}
\alias{lisaTest}
\title{Local indicators of spatial association}
\usage{
lisaTest(x, weight, na.rm = FALSE, alternative = "greater")
}
\arguments{
\item{x}{Feature value}

\item{weight}{Adjacency matrix}

\item{na.rm}{Whether to remove NAs from feature value}

\item{alternative}{"two.sided", "less", or "greater"}
}
\description{
Local indicators of spatial association
}
