% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data (client-10-245-74-74.fas.harvard.edu's
%   conflicted copy 2019-11-09).R, R/data.R
\docType{data}
\name{mOB}
\alias{mOB}
\title{Mouse olfactory bulb spatial transcriptomics transcriptomics}
\format{List where 'counts' is a sparse matrix with columns as voxels and rows as genes and
                   'pos' is a data frame of x and y position values per voxel}
\source{
\url{https://science.sciencemag.org/content/353/6294/78}

\url{http://www.spatialtranscriptomicsresearch.org/datasets/doi-10-1126science-aaf2403/}
}
\usage{
mOB

mOB
}
\description{
Mouse olfactory bulb spatial transcriptomics transcriptomics

Mouse olfactory bulb spatial transcriptomics transcriptomics
}
\keyword{datasets}
