% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationStatistics.R
\name{moranPermutationTest}
\alias{moranPermutationTest}
\title{Spatialy autocorrelation by Moran's I using permutation testing}
\usage{
moranPermutationTest(z, w, na.rm = FALSE, alternative = "two.sided",
  N = 10000, seed = 0, ncores = 1, plot = FALSE, ...)
}
\arguments{
\item{z}{Feature value}

\item{w}{Adjacency matrix}

\item{na.rm}{Whether to remove NAs from feature value}

\item{alternative}{"two.sided", "less", or "greater"}

\item{N}{Number of permutations}

\item{seed}{Random seed}

\item{plot}{Plot permutated distribution}

\item{...}{Additional parameters to pass to histogram plotting}

\item{n.cores}{Number of cores for parallel processing}
}
\description{
Spatialy autocorrelation by Moran's I using permutation testing
}
