% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationStatistics.R
\name{moranTest}
\alias{moranTest}
\title{Moran's I compute from scratch
x is value
weight is adjacency matrix (weights)
use log.p since many p-values very small (avoid 0s)}
\usage{
moranTest(x, weight, na.rm = FALSE, alternative = "greater")
}
\description{
Moran's I compute from scratch
x is value
weight is adjacency matrix (weights)
use log.p since many p-values very small (avoid 0s)
}
