% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouseCoronal}
\alias{mouseCoronal}
\title{Visium 10X Spatial Transcriptomics data of an adult mouse brain coronal section (P56)}
\format{
List where 'filteredGenes' is a sparse matrix with columns as genes and rows as spot IDs
                         where expression values have already been normalized and genes filtered
                         to 1263 genes whose expression variance across the 2702 spots is higher
                         than transcriptome-wide expectations. mt genes also removed.
                   'tissueSpotRotation' is a data frame of x, y position values per spot
                         where the spot barcodes are the 2702 that overlap the tissue.
                         Coordinated have been adjusted to match the high resolution PNG image
                         and rotated such that they visually match the tissue image when plotted.
                   'cluster6CorrMtx' is the cross correlation matrix of the significantly spatially
                         variable expressed genes in cluster 6.
}
\source{
\url{https://cf.10xgenomics.com/samples/spatial-exp/1.1.0/V1_Adult_Mouse_Brain/}
}
\usage{
mouseCoronal
}
\description{
Visium 10X Spatial Transcriptomics data of an adult mouse brain coronal section (P56)
}
\keyword{datasets}
