% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotNeighborBoxplot}
\alias{plotNeighborBoxplot}
\title{plot boxplot of expression for nearest neighbor}
\usage{
plotNeighborBoxplot(gexpA, gexpB, groupA, groupB, weight)
}
\description{
plot boxplot of expression for nearest neighbor
}
