% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pos}
\alias{pos}
\title{MOB spatial transcriptomics positional data}
\format{Data frame of x and y position values per voxel}
\source{
\url{http://www.spatialtranscriptomicsresearch.org/datasets/doi-10-1126science-aaf2403/}
}
\usage{
pos
}
\description{
MOB spatial transcriptomics positional data
}
\keyword{datasets}
