% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{purkinje}
\alias{purkinje}
\title{SlideSeq data of the Purkinje layer of the mouse cerebellum for one puck (Puck_180819_12)}
\format{
List where 'counts' is a sparse matrix with columns as voxels and rows as genes and
                   'pos' is a data frame of x and y position values per voxel
                         and slice index for 4 consecutive slices
}
\source{
\url{https://science.sciencemag.org/content/363/6434/1463}
}
\usage{
purkinje
}
\description{
SlideSeq data of the Purkinje layer of the mouse cerebellum for one puck (Puck_180819_12)
}
\keyword{datasets}
