% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationStatistics.R
\name{signedLisa}
\alias{signedLisa}
\title{Signed local indicators of spatial association to identify regions driving global autocorrelation}
\usage{
signedLisa(x, weight, alternative = "greater", ...)
}
\arguments{
\item{x}{Feature value}

\item{weight}{Adjacency weight matrix}

\item{alternative}{"two.sided", "less", or "greater"}

\item{...}{Additional plotting parameters}

\item{pos}{Position}

\item{plot}{Boolean of whether to plot}
}
\value{
Signed LISA statistic for each point
}
\description{
Signed local indicators of spatial association to identify regions driving global autocorrelation
}
\examples{
data(mOB)
pos <- mOB$pos
weight <- voronoiAdjacency(pos)
gexp <- normalizeCounts(mOB$counts, log=FALSE, verbose=FALSE)['Camk4',]
plotEmbedding(pos, colors=scale(gexp)[,1], cex=3)
lisa <- signedLisa(gexp, weight, cex=3)
plotEmbedding(pos, colors=slisa, gradientPalette=colorRampPalette(c('darkgreen', 'white', 'darkorange'))(100))

}
