% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExpressionData_object.R
\docType{data}
\name{SpatialExpressionData}
\alias{SpatialExpressionData}
\title{R6 object for storing spatial expression data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SpatialExpressionData
}
\description{
R6 object for storing spatial expression data
}
\keyword{datasets}
