% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestNeighbor.R
\name{voronoiAdjacency}
\alias{voronoiAdjacency}
\title{Neighbor weight matrix by voronoi adjacency}
\usage{
voronoiAdjacency(pos, filterDist = NA, plot = FALSE)
}
\arguments{
\item{pos}{Position}

\item{filterDist}{Euclidean distance beyond which two cells cannot be considered neighbors}

\item{plot}{Whether to plot}
}
\description{
Neighbor weight matrix by voronoi adjacency. Modified from Barry Rowlingson and Alison Hale's voronoi_adjacency function in the caramellar package
}
\examples{
{
data(mOB)
pos <- mOB$pos
w <- voronoiAdjacency(pos, plot=TRUE)
}

}
