% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data (client-10-245-74-74.fas.harvard.edu's
%   conflicted copy 2019-11-09).R, R/data.R
\docType{data}
\name{BCL3}
\alias{BCL3}
\title{Breast cancer biopsy layer 3 bulb spatial transcriptomics transcriptomics}
\format{List where 'counts' is a sparse matrix with columns as voxels and rows as genes and
                   'pos' is a data frame of x and y position values per voxel}
\source{
\url{https://science.sciencemag.org/content/353/6294/78}

\url{https://science.sciencemag.org/content/353/6294/78}
}
\usage{
BCL3

BCL3
}
\description{
Breast cancer biopsy layer 3 bulb spatial transcriptomics transcriptomics

Breast cancer biopsy layer 3 bulb spatial transcriptomics transcriptomics
}
\keyword{datasets}
