% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drosophila}
\alias{drosophila}
\title{Drosophila embryo aligned ISH}
\format{An object of class \code{list} of length 4.}
\usage{
drosophila
}
\description{
Drosophila embryo aligned ISH
}
\keyword{datasets}
