% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filterSpatialPatterns}
\alias{filterSpatialPatterns}
\title{Filter for spatial patterns}
\usage{
filterSpatialPatterns(
  mat,
  I,
  w,
  adjustPv = TRUE,
  alpha = 0.05,
  minPercentCells = 0.05,
  verbose = TRUE,
  details = FALSE
)
}
\arguments{
\item{mat}{Gene expression matrix.}

\item{I}{Output of getSpatialPatterns}

\item{w}{Weight adjacency matrix}

\item{adjustPv}{Whether to perform BH multiple testing correction}

\item{alpha}{P-value threshold for LISA score to be considered significant.}

\item{minPercentCells}{Minimum percent of cells that must be driving spatial pattern}

\item{verbose}{Verbosity}

\item{details}{Return details}
}
\description{
Filter out spatial patterns driven by small number of cells using LISA
}
