% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestNeighbor.R
\name{getKnn}
\alias{getKnn}
\title{K nearest neighbors}
\usage{
getKnn(pos, k)
}
\arguments{
\item{pos}{Position}

\item{k}{Number of nearest neighbors}
}
\description{
K nearest neighbors in space based on position
}
