% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getSpatialPatterns}
\alias{getSpatialPatterns}
\title{Identify spatial clusters}
\usage{
getSpatialPatterns(mat, weight, alternative = "greater",
  verbose = TRUE)
}
\arguments{
\item{mat}{Gene expression matrix. Must be normalized such that correlations
will not be driven by technical artifacts.}

\item{weight}{Spatial weights such as a weighted adjacency matrix}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "greater" (default), "two.sided" or "less".}

\item{verbose}{Verbosity}
}
\description{
Identify spatially clustered genes using Moran's I
}
