% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{groupSigSpatialPatterns}
\alias{groupSigSpatialPatterns}
\title{Group significant spatial patterns}
\usage{
groupSigSpatialPatterns(pos, mat, d, hclustMethod = "complete", trim = 0,
  deepSplit = 0, minClusterSize = 0, plot = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{pos}{Position matrix where each row is a cell, columns are
x, y, (optionally z) coordinations}

\item{mat}{Gene expression matrix. Must be normalized such that correlations
will not be driven by technical artifacts.}

\item{d}{Distance matrix relating genes. Spatial cross-correlation recommended.}

\item{hclustMethod}{Method for hclust()}

\item{deepSplit}{Tuning parameter for dynamic tree cutting cutreeDynamic()}

\item{minClusterSize}{Smallest gene cluster size}
}
\description{
Identify primary spatial patterns using hierarchical clustering and dynamic tree cutting
}
