% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{interpolate}
\alias{interpolate}
\title{Gridded bivariate interpolation}
\usage{
interpolate(pos, gexp, trim = 0, zlim = c(-1, 1), fill = TRUE,
  binSize = 100, cex = 1, col = colorRampPalette(c("blue", "white",
  "red"))(100), plot = TRUE, ...)
}
\description{
Gridded bivariate interpolation
}
