% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{map2col}
\alias{map2col}
\title{Helper function to map values to colors
Source: https://stackoverflow.com/questions/15006211/how-do-i-generate-a-mapping-from-numbers-to-colors-in-r}
\usage{
map2col(x, pal = colorRampPalette(c("blue", "grey", "red"))(100),
  na.col = "lightgrey", limits = NULL)
}
\description{
Helper function to map values to colors
Source: https://stackoverflow.com/questions/15006211/how-do-i-generate-a-mapping-from-numbers-to-colors-in-r
}
