% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationStatistics.R
\name{moranPermutationTest}
\alias{moranPermutationTest}
\title{Permutation test to assess for significance of Moran's I}
\usage{
moranPermutationTest(z, w, na.rm = FALSE, alternative = "two.sided",
  N = 10000, seed = 0, ncores = parallel::detectCores() - 1,
  plot = FALSE, ...)
}
\description{
Permutation test to assess for significance of Moran's I
}
