% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{moranTest}
\alias{moranTest}
\title{Moran's I compute from scratch
x is value
w is adjacency matrix (weights)}
\usage{
moranTest(x, w)
}
\description{
Moran's I compute from scratch
x is value
w is adjacency matrix (weights)
}
