% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Plot neighbor network
https://stackoverflow.com/questions/43879347/plotting-a-adjacency-matrix-using-pure-r}
\usage{
plotNetwork(pos, adj, line.col = "red", line.power = 1, ...)
}
\description{
Plot neighbor network
https://stackoverflow.com/questions/43879347/plotting-a-adjacency-matrix-using-pure-r
}
