% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data (client-10-245-74-74.fas.harvard.edu's
%   conflicted copy 2019-11-09).R
\docType{data}
\name{purkinje}
\alias{purkinje}
\title{SlideSeq data}
\format{An object of class \code{list} of length 2.}
\usage{
purkinje
}
\description{
SlideSeq data
}
\keyword{datasets}
