% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationStatistics.R
\name{spatialCrossCor}
\alias{spatialCrossCor}
\title{inter-cell-type cross cor}
\usage{
spatialCrossCor(gexpA, gexpB, groupA, groupB, weight = NULL, pos = NULL,
  k = 3)
}
\description{
inter-cell-type cross cor
}
