% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationStatistics
%   (client-10-245-74-74.fas.harvard.edu's conflicted copy 2019-11-09).R,
%   R/correlationStatistics.R
\name{spatialCrossCor}
\alias{spatialCrossCor}
\title{Calculates spatial cross correlation for two features}
\usage{
spatialCrossCor(x, y, weight)

spatialCrossCor(x, y, weight)
}
\arguments{
\item{x}{Feature 1 value}

\item{y}{Feature 2 value}

\item{weight}{Adjacency weight matrix}

\item{x}{Feature 1 value}

\item{y}{Feature 2 value}

\item{weight}{Adjacency weight matrix}
}
\value{
Spatial cross correlation statistic

Spatial cross correlation statistic
}
\description{
Calculates spatial cross correlation for two features

Calculates spatial cross correlation for two features
}
\examples{
data(mOB)
pos <- mOB$pos
weight <- getSpatialNeighbors(pos)
gexp <- normalizeCounts(mOB$counts, log=FALSE, verbose=FALSE)
scc <- spatialCrossCor(gexp[1,], gexp[2,], weight)

data(mOB)
pos <- mOB$pos
weight <- getSpatialNeighbors(pos)
gexp <- normalizeCounts(mOB$counts, log=FALSE, verbose=FALSE)
scc <- spatialCrossCor(gexp[1,], gexp[2,], weight)

}
