% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationStatistics.R
\name{spatialCrossCorMatrix}
\alias{spatialCrossCorMatrix}
\title{Calculates a spatial cross correlation matrix}
\usage{
spatialCrossCorMatrix(sigMat, w)
}
\description{
Calculates a spatial cross correlation matrix
}
