% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationStatistics
%   (client-10-245-74-74.fas.harvard.edu's conflicted copy 2019-11-09).R,
%   R/correlationStatistics.R
\name{spatialCrossCorTest}
\alias{spatialCrossCorTest}
\title{Tests for significance of spatial cross correlation for two features using random label null model}
\usage{
spatialCrossCorTest(x, y, w, n = 1000, plot = FALSE, ...)

spatialCrossCorTest(x, y, w, n = 1000, plot = FALSE, ...)
}
\arguments{
\item{x}{Feature 1 value}

\item{y}{Feature 2 value}

\item{w}{Binary weight matrix}

\item{n}{Permutation iterations}

\item{plot}{Plot permutated distribution}

\item{...}{Additional parameters to pass to histogram plotting}

\item{x}{Feature 1 value}

\item{y}{Feature 2 value}

\item{w}{Binary weight matrix}

\item{n}{Permutation iterations}

\item{plot}{Plot permutated distribution}

\item{...}{Additional parameters to pass to histogram plotting}
}
\value{
Two-sided test p-value

Two-sided test p-value
}
\description{
Tests for significance of spatial cross correlation for two features using random label null model

Tests for significance of spatial cross correlation for two features using random label null model
}
\examples{
\dontrun{
data(mOB)
pos <- mOB$pos
w <- getSpatialNeighbors(pos, plot=FALSE)
gexp <- normalizeCounts(mOB$counts, log=FALSE, verbose=FALSE)
pv1 <- spatialCrossCorTest(gexp['Gpsm1',], gexp['Nrgn',], w)
pv2 <- spatialCrossCorTest(gexp['Gpsm1',], gexp['Glul',], w)
pv3 <- spatialCrossCorTest(gexp[1,], gexp[2,], w)
}
\dontrun{
data(mOB)
pos <- mOB$pos
w <- voronoiAdjacency(pos, plot=FALSE)
gexp <- normalizeCounts(mOB$counts, log=FALSE, verbose=FALSE)
pv1 <- spatialCrossCorTest(gexp['Gpsm1',], gexp['Nrgn',], w)
pv2 <- spatialCrossCorTest(gexp['Gpsm1',], gexp['Glul',], w)
pv3 <- spatialCrossCorTest(gexp[1,], gexp[2,], w)
}
}
