% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestNeighbor.R
\name{voronoiAdjacency}
\alias{voronoiAdjacency}
\title{Neighbor weight matrix by voronoi adjacency}
\usage{
voronoiAdjacency(pos, njitter = 0, ajitter = 0, filterDist = NA,
  plot = FALSE)
}
\arguments{
\item{pos}{Position}

\item{njitter}{Number of times to jitter positions (useful for gridded spatial transcriptomics data)}

\item{ajitter}{Amount to jitter}

\item{filterDist}{Euclidean distance beyond which two cells cannot be considered neighbors}

\item{plot}{Whether to plot}
}
\description{
Neighbor weight matrix by voronoi adjacency. Modified from Barry Rowlingson and Alison Hale's voronoi_adjacency function in the caramellar package
}
\examples{
{
data(mOB)
pos <- mOB$pos
w <- voronoiAdjacency(pos, njitter=100, ajitter=5, filterDist=1.8)
}

}
